refactor(functions): refactor function, move logic out of the main file
refactored function;
moved logic out of the main file;
